requirements.txt
boto3==1.28.53
boto3==1.26.76
elasticsearch==7.10.0
neo4j==5.14.1
pandas==2.0.0
python-dotenv==1.0.0
requests_aws4auth==1.1.1
spacy==3.7.2
